Update devEtherIP.c
Changed wf_add_record so that it doesn't use the NELM field as the Count argument as that was making a CIP request that the PLC didn't understand. 

Removed the check for Element count from lock_data as our asked for elements is only 1 but we receive the whole array back. 

In wf_read I removed the comparison of tag->elements and rec->nelm as out tag elements is only 1 when asking for a whole array.